Fixed Race Condition Bug with File Upload
We should store the file on the firebase server before
updating the database to prevent a race condition between
concurrent clients. Since clients will randomly receive
a file from the database, if we update the database then
get the file from firebase storage, we risk trying to
get a file that has not been stored on firebase.